refactor: 컨설팅 답변 tab 컴포넌트 widget으로 폴더 이동
